[CI] Add planet scene validation job
