+++ Space Game/Assets/Scripts/Editor/PlanetSceneValidator.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 81f1c1ad6e9449f09d4770c8141bcbb8
