+++ Space Game/Assets/Tests/EditMode/Editor/FriendSlop.EditModeTests.asmdef
@@ -2,6 +2,7 @@
     "name": "FriendSlop.EditModeTests",
     "rootNamespace": "FriendSlop.Tests.EditMode",
     "references": [
+        "FriendSlop.Editor",
         "FriendSlop.Runtime",
         "Unity.Netcode.Runtime",
         "Unity.Services.Core",
