+++ Space Game/Assets/Tests/EditMode/Editor/PlanetSceneValidatorTests.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 8ff2f6a245034d98b7995e70ae1ef0a3
